+++ commands/quarter.js
@@ -1,43 +1,13 @@
-const fs = require('fs')
-const path = require('path')
-const mkdirp = require('mkdirp')
-const subQuarters = require('date-fns/sub_quarters')
-const addQuarters = require('date-fns/add_quarters')
-
-const getYear = require('../lib/get-year')
-const getQuarter = require('../lib/get-quarter')
-const createQuarterFile = require('../lib/create-quarter-file')
+const quarter = require('../lib/quarter')
+const createFile = require('../lib/create-file')
 
 function command (args, flags, context) {
-  let date = new Date()
-
-  if (flags.next) {
-    date = addQuarters(date, 1)
-  } else if (flags.previous) {
-    date = subQuarters(date, 1)
-  }
-
-  const quarter = getQuarter(args.quarter, date)
-
-  let dateForYear = date
-  if (quarter.number === 1) {
-    dateForYear = quarter.end
-  }
-
-  const year = getYear(args.year, dateForYear)
-
-  const quarterFile = createQuarterFile({ quarter, year })
-
-  const quarterDirectoryPath = path.join(process.cwd(), year.number, 'quarters')
-  const quarterFilepath = path.join(quarterDirectoryPath, `${quarter.name.toLowerCase()}.md`)
-
-  mkdirp.sync(quarterDirectoryPath)
-
-  try {
-    fs.accessSync(quarterFilepath)
-  } catch (e) {
-    fs.writeFileSync(quarterFilepath, quarterFile)
-  }
+  createFile(quarter({
+    next: flags.next,
+    previous: flags.previous,
+    quarter: args.quarter,
+    year: args.year
+  }))
 }
 
 const args = [