+++ commands/week.js
@@ -1,42 +1,13 @@
-const fs = require('fs')
-const path = require('path')
-const mkdirp = require('mkdirp')
-const subWeeks = require('date-fns/sub_weeks')
-const addWeeks = require('date-fns/add_weeks')
-
-const getYear = require('../lib/get-year')
-const getWeek = require('../lib/get-week')
-const createWeekFile = require('../lib/create-week-file')
+const week = require('../lib/week')
+const createFile = require('../lib/create-file')
 
 function command (args, flags, context) {
-  let date = new Date()
-
-  if (flags.next) {
-    date = addWeeks(date, 1)
-  } else if (flags.previous) {
-    date = subWeeks(date, 1)
-  }
-  const week = getWeek(args.week, date)
-
-  let dateForYear = date
-  if (week.number === 1) {
-    dateForYear = week.end
-  }
-
-  const year = getYear(args.year, dateForYear)
-
-  const weekFile = createWeekFile({ week, year })
-
-  const weekDirectoryPath = path.join(process.cwd(), year.number, 'weeks')
-  const weekFilepath = path.join(weekDirectoryPath, `${week.numberZeroFilled}.md`)
-
-  mkdirp.sync(weekDirectoryPath)
-
-  try {
-    fs.accessSync(weekFilepath)
-  } catch (e) {
-    fs.writeFileSync(weekFilepath, weekFile)
-  }
+  createFile(week({
+    next: flags.next,
+    previous: flags.previous,
+    week: args.week,
+    year: args.year
+  }))
 }
 
 const args = [